fix display proposals
